Allow separate 'integration tests'
Existing tests are pure, and run against a dummy, in-memory alternative to Yabai

We want to make sure this dummy acts like the real Yabai, which we can do using
integration tests. These would set up Yabai and our dummy to the same state, run
a command or query against both, and check that the result is the same. Since
this requires interacting with the real window manager, we don't want to do it
as part of the normal testing process, so we give it an env var flag.